requirements.txt
email-validator==1.1.3
Flask==2.1.0
Jinja2==3.0.1
langchain==0.0.123
names==0.3.0
openai>=0.27.2
git+https://github.com/mmabrouk/chatgpt-wrapper@main
openpyxl==3.0.7
playwright==1.13.1
prompt-toolkit==3.0.20
pydantic-computed==0.1.1
pyperclip==1.8.2
python-frontmatter==1.0.0
PyYAML==5.4.1
rich==10.2.2
sqlalchemy==1.4.26
tiktoken
fastapi==0.70.0
uvicorn==0.15.0
python-multipart==0.0.5
pymongo==4.0.1